Restructure app and move code into lib crate- move nearly all code out of main.rs- move ui_state::UiState -> ui::state::State
